% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoinToss.R
\name{calc_toss_probs}
\alias{calc_toss_probs}
\title{Title: Probability calculator (All Heads and Tails for Now).}
\usage{
calc_toss_probs(num_coins = NULL, probability = c("all_heads", "all_tails"))
}
\arguments{
\item{num_coins}{The number of coins tossed.}

\item{probability}{The probability you intend to calculate.
For now 'all-heads' and 'all_tails'.}
}
\value{
A numeric value for the calculated probability.
}
\description{
Title: Probability calculator (All Heads and Tails for Now).
}
\examples{
all_heads_p = calc_toss_probs(3)
}
