% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coin.R
\name{coin}
\alias{coin}
\alias{is.coin}
\title{Coin}
\usage{
coin(sides = c("heads", "tails"), prob = c(0.5, 0.5))

is.coin(x)
}
\arguments{
\item{sides}{vector of coin sides}

\item{prob}{vector of side probabilities}

\item{x}{an R object}
}
\value{
an object of class coin
}
\description{
Creates an object of class \code{"coin"}
}
\examples{
# default 
coin1 <- coin()

# another coin
coin2 <- coin(c('h', 't'))

# us cent
cent1 <- coin(c('lincoln', 'shield'))

# loaded coin
loaded <- coin(prob = c(0.7, 0.3))

}

