% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toss.R
\name{make_toss}
\alias{make_toss}
\title{Make Toss Object}
\usage{
make_toss(coin, flips)
}
\arguments{
\item{coin}{object of class coin}

\item{flips}{object of class flips}
}
\description{
Constructor function for object "toss"
}
\keyword{internal}

