% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-toss.R
\name{plot.toss}
\alias{plot.toss}
\title{Plot of object toss}
\usage{
\method{plot}{toss}(x, side = 1, ...)
}
\arguments{
\item{x}{an object of class \code{"coin"}}

\item{side}{number indicating which side of coin to consider}

\item{\dots}{arguments to be passed to/from other methods}
}
\description{
Plots the relative frequencies of a series of tosses
}
\examples{
 \dontrun{
 # create a coin and toss it 100 times
 coin1 <- coin()
 toss1000 <- toss(coin1, times = 1000)
 
 plot(toss1000)
 }
}

