% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toss.R
\name{toss}
\alias{as.toss}
\alias{is.toss}
\alias{toss}
\title{Toss a coin}
\usage{
toss(coin, times = 1)

is.toss(x)

as.toss(x)
}
\arguments{
\item{coin}{object of class \code{"coin"}}

\item{times}{number of tosses}

\item{x}{an R object}
}
\value{
an object of class \code{"toss"} with the following elements:

\item{tosses}{vector of tosses}

\item{sides}{vector of coin \code{"sides"}}

\item{prob}{vector of coin \code{"prob"}}

\item{total}{total number of tosses}

\item{heads}{number of heads}

\item{tails}{number of tails}
}
\description{
Creates an object of class \code{"toss"}
}
\examples{
coin1 <- coin()

# toss a coin 10 times
toss10 <- toss(coin1, times = 10)

# add 5 more tosses
toss15 <- toss10 + 5
}

