% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoinToss.R
\name{tossCoins}
\alias{tossCoins}
\title{Title: Coins Sampling.}
\usage{
tossCoins(coins, output = c("char", "num", "list"))
}
\arguments{
\item{coins}{The coins population which is a no parameter subordinate function.}

\item{output}{A data frame with numeric or character outcomes}
}
\value{
A data frame showing the coins samples generated.
}
\description{
Title: Coins Sampling.
}
\note{
Each row represents a sample.
}
\examples{
coin_perms <- tossCoins(coinsPop(3), 'list')
}
