% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoinToss.R
\name{coinsPop}
\alias{coinsPop}
\title{Title: Coin population.}
\usage{
coinsPop()
}
\value{
The number of coins tossed represented by a string of Heads and Tails.
}
\description{
Title: Coin population.
}
\note{
Multiple coin tosses is a situation where the population is less than the sample.
That means 2 coins (considered the population) yield 4 samples; counter intuitive right.
}
\examples{
coins_population <- coinsPop()
}
